+++ tags/day.tag
@@ -1,30 +1,21 @@
 <day>
-  <div class="two columns" if={day > 25}>
-    <p>{ day }일</p>
-  </div>
-  <div class="two columns" if={day < 26}>
-    <p if={ isEmpty() }>
-      { day }일
-      <br>
-      비어있습니다.
-      <br>
-      <a href="#" if={ !formOpened } onclick={ openNewForm }>등록하기</a>
-    </p>
+  <div class="cell-inner">
+    <div class="cell-header">{ day }</div>
 
-    <p if={ !isEmpty() }>
-      { day }일:
-      <br>
-      { author }님의 <a href={url} target="_blank">{title}</a>
-      <br>
-      <span if={ isOwned() }>
-        <a href="#" if={ !formOpened } onclick={ editForm }>고치기</a>
-        <a href="#" if={ !formOpened } onclick={ delete }>취소하기</a>
-      </span>
-    </p>
-
-    <br>
-
+    <div class="cell-body" if={ day > 25 }></div>
+    <div class="cell-body" if={ day < 26 && isEmpty() }>
+      <a href="#" if={ !formOpened } onclick={ openNewForm }>예약하기</a>
+    </div>
+    <div class="cell-body" if={ day < 26 && !isEmpty() }>
+        { author }님<br>
+        <a href={url} target="_blank">{title}</a><br>
+        <span if={ isOwned() }>
+          <a href="#" if={ !formOpened } onclick={ editForm }>고치기</a>
+          <a href="#" if={ !formOpened } onclick={ delete }>취소하기</a>
+        </span>
+    </div>
     <div if={ formOpened }>
+      <br>
       <input name="day[id]" type="hidden" value={ day }>
       <input name="day[author]" type="text" placeholder="작성자 이름">
       <input name="day[title]" type="text" placeholder="제목">
@@ -57,6 +48,7 @@
     }
 
     delete() {
+      if(!confirm("정말로 취소하시겠어요?")) { return }
       this.parent.parent.opts.deleteDay(this.day)
       this.parent.parent.opts.loadData()
     }
@@ -66,7 +58,7 @@
     }
 
     openNewForm() {
-      if(!this.parent.uid) { alert("Please log-in"); return }
+      if(!this.parent.uid) { alert("로그인해주세요."); return }
       this.formOpened = true
     }
